Fixed some BuddyAllocator::Request() traversal issues
If a Request(size) wants a slot on level n where n < max_levels, but no slots on level n are free, it would continue traversing on level n + 1 instead and find the first free slot, even if the slot is too small.

On some situations, it didn't traverse back to the parent when both buddies of a parent were not free.